requirements.txt
boto3==1.9.215
botocore==1.12.215
Django==2.1.5
django-extensions==2.2.1
django-shell-plus==1.1.7
djangorestframework==3.9.1
docutils==0.15.2
jmespath==0.9.4
psycopg2==2.7.7
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
